Windows warning fix
git-svn-id: https://svn.osgeo.org/gdal/trunk@40312 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/port/cpl_error.cpp
@@ -591,7 +591,7 @@ void CPLDebug( const char * pszCategory,
         struct CPLTimeVal tv;
         CPLGettimeofday(&tv, NULL);
         strcpy( pszMessage, "[" );
-        strcat( pszMessage, VSICTime( tv.tv_sec ) );
+        strcat( pszMessage, VSICTime( static_cast<unsigned long>(tv.tv_sec) ) );
 
         // On windows anyway, ctime puts a \n at the end, but I'm not
         // convinced this is standard behaviour, so we'll get rid of it
